[#82] Refactor : enum Condition 패키지 수정
-  사용하는 패키지에 위치하도록 수정
